Remove single-use proxy abstractions

Copilot-Session: 239ec5f7-870a-4259-bcae-4ca85fb913a0
